requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
filter==0.0.0.20200724
Markdown==3.3.7
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
